Fix Linux audio command
`Open3.popen3` requires the command in one of a few forms:

- Arguments of: command + arguments
- An array of: command + arguments
- A string of command + arguments

Previously, with multiple arguments, it was using the first form, and so trying to run an executable that has the name `command -v mplayer >/dev/null 2>&1 && mplayer -msglevel all=-1 -nolirc`, which obviously doesn't exist (and providing one command-line argument, being the audio file `path`).

Hence this error prevented startup of Snek:

```
➜ (cd ../snek && bundle _1.17.3_ exec ./bin/snek)
bundler: failed to load command: ./bin/snek (./bin/snek)
Errno::ENOENT: No such file or directory - command -v mplayer >/dev/null 2>&1 && mplayer -msglevel all=-1 -nolirc
  /home/brendan/.rbenv/versions/2.7.2/lib/ruby/2.7.0/open3.rb:213:in `spawn'
  /home/brendan/.rbenv/versions/2.7.2/lib/ruby/2.7.0/open3.rb:213:in `popen_run'
  /home/brendan/.rbenv/versions/2.7.2/lib/ruby/2.7.0/open3.rb:101:in `popen3'
  /home/brendan/.rbenv/versions/2.7.2/lib/ruby/gems/2.7.0/gems/terminal_game_engine-0.1.2/lib/terminal_game_engine/sound.rb:10:in `play'
  /home/brendan/Projects/snek/lib/game.rb:278:in `play_sound'
  /home/brendan/Projects/snek/lib/game.rb:21:in `start'
  /home/brendan/Projects/snek/bin/snek:19:in `<top (required)>'
```

Converting the command into one of the first two forms doesn't work, as the `command` command is actually a shell built-in command (so not available outside of a shell) and only the third form uses a shell to evaluate the given command. Thus, we need to use shellwords to safely escape the path.